chore: move ui components to toplevel
