+++ ui/src/app/console/Console.tsx
@@ -20,7 +20,7 @@ import Dropdown from '~/components/Dropdown';
 import EmptyState from '~/components/EmptyState';
 import Input from '~/components/forms/Input';
 import { JsonEditor } from '~/components/json/JsonEditor';
-import { PageHeader } from '~/components/ui/page';
+import { PageHeader } from '~/components/Page';
 
 import { IAuthMethod } from '~/types/Auth';
 import { Command } from '~/types/Cli';
