+++ ui/src/app/flags/Flag.tsx
@@ -16,7 +16,7 @@ import MoreInfo from '~/components/MoreInfo';
 import FlagForm from '~/components/flags/FlagForm';
 import CopyToNamespacePanel from '~/components/panels/CopyToNamespacePanel';
 import DeletePanel from '~/components/panels/DeletePanel';
-import { PageHeader } from '~/components/ui/page';
+import { PageHeader } from '~/components/Page';
 
 import { useError } from '~/data/hooks/error';
 import { useSuccess } from '~/data/hooks/success';
