+++ ui/src/components/TableSkeleton.tsx
@@ -1,4 +1,4 @@
-import { Skeleton } from '~/components/ui/skeleton';
+import { Skeleton } from '~/components/Skeleton';
 
 const TableSkeleton = () => {
   return (
