+++ ui/src/components/flags/FlagTable.tsx
@@ -21,9 +21,9 @@ import {
 import { Badge } from '~/components/Badge';
 import Searchbox from '~/components/Searchbox';
 import Well from '~/components/Well';
-import { DataTablePagination } from '~/components/ui/table-pagination';
-import { TableSkeleton } from '~/components/ui/table-skeleton';
-import { DataTableViewOptions } from '~/components/ui/table-view-options';
+import { DataTablePagination } from '~/components/TablePagination';
+import { TableSkeleton } from '~/components/TableSkeleton';
+import { DataTableViewOptions } from '~/components/TableViewOptions';
 
 import { IEnvironment } from '~/types/Environment';
 import { FlagType, IFlag, flagTypeToLabel } from '~/types/Flag';
